+++ lib/router.dart
@@ -3,6 +3,8 @@ import 'package:go_router/go_router.dart';
 import 'package:simple_design/screens/home/home.screen.dart';
 import 'package:simple_design/screens/buttons/buttons.screen.dart';
 import 'package:simple_design/screens/forms/forms.screen.dart';
+import 'package:simple_design/screens/cards/cards.screen.dart';
+import 'package:simple_design/screens/data_display/data_display.screen.dart';
 import 'package:simple_design/widgets/showcase_shell.dart';
 
 final GlobalKey<NavigatorState> globalNavigatorKey = GlobalKey<NavigatorState>();
@@ -31,6 +33,18 @@ final router = GoRouter(
           GoRoute(
             path: FormsScreen.routeName,
             builder: (context, state) => const FormsScreen(),
+          ),
+        ]),
+        StatefulShellBranch(routes: [
+          GoRoute(
+            path: CardsScreen.routeName,
+            builder: (context, state) => const CardsScreen(),
+          ),
+        ]),
+        StatefulShellBranch(routes: [
+          GoRoute(
+            path: DataDisplayScreen.routeName,
+            builder: (context, state) => const DataDisplayScreen(),
           ),
         ]),
       ],